pyproject.toml
[tool.poetry]
name = "ted"
version = "0.1.0"
description = ""
authors = ["alexthejpeg <alex85120@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.6.0"
Pillow = "^8.1.0"
requests = "^2.25.1"
googletrans = "4.0.0rc1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
flake8-bugbear = "^20.11.1"
black = "^20.8b1"
isort = "^5.7.0"
flake8-isort = "^4.0.0"

[tool.poetry.scripts]
ted = "ted.bot:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
